feat: added refresh token and cleaned up some logic

+++ nanolink_api/src/features/links/use_cases/delete_link_use_case.dart
@@ -1,20 +1,21 @@
 import 'package:nanolink_core/nanolink_core.dart';
-import 'package:supabase/supabase.dart';
+import 'package:supabase/supabase.dart' hide User;
 
 class DeleteLinkUseCase {
   DeleteLinkUseCase(this._client);
 
   final SupabaseClient _client;
 
-  Future<LinkDto> run(int id) async {
+  Future<LinkDto> run(int id, User user) async {
     final response = await _client
         .from(LinkDto.tableName)
         .delete()
         .eq('id', id)
+        .eq('user_id', user.id)
         .maybeSingle()
         .execute()
         .catchError(
-          (_) => throw const LinksException.deleteLinkException(),
+          (_) => throw const LinksException.deleteLink(),
         );
     final data = response.data as Map<String, dynamic>?;
 
@@ -24,10 +25,10 @@ class DeleteLinkUseCase {
 
         return link;
       } catch (_) {
-        throw const LinksException.linkDeserializationException();
+        throw const LinksException.linkDeserialization();
       }
     } else {
-      throw const LinksException.deleteLinkException();
+      throw const LinksException.deleteLink();
     }
   }
 }
